requirements.txt
Django==3.1.2
psycopg2
django-phonenumber-field
djangorestframework
phonenumbers
PyJWT
